Shahzad suggestions
